+++ Types/MarkX/get_Table.ps1
@@ -1,1 +1,9 @@
-$this.XML | Select-Xml -XPath '//table' | Select-Object -ExpandProperty Node
+<#
+.SYNOPSIS
+    Gets any tables
+.DESCRIPTION
+    Gets any tables present in the markdown
+#>
+$this.XML | 
+    Select-Xml -XPath '//table' | 
+    Select-Object -ExpandProperty Node